+++ db/vector/__init__.py
@@ -2,13 +2,13 @@
 
 from .milvus_store import init_milvus
 from .vector_query import search
-from .rag_pipeline import rag_answer, qwen_generate
+from .rag_pipeline import rag_answer, llm_generate
 from .build_vector_db import populate
 
 __all__ = [
     "init_milvus",
     "search",
     "rag_answer",
-    "qwen_generate",
+    "llm_generate",
     "populate",
 ]